Extracted EmptyView from the KanbanList
