+++ src/_bulmaVariables.scss
@@ -1,1 +1,1 @@
-@import "~bulma/sass/utilities/initial-variables";
+@import "~bulma/sass/utilities/mixins";
