+++ src/templates/KanbanBoard/EmptyView.module.scss
@@ -1,0 +1,10 @@
+@import '../../bulmaVariables';
+
+.emptyContainer {
+  width: 100%;
+  height: 100%;
+}
+
+.emptyContent {
+  margin: $size-normal;
+}
