+++ src/templates/KanbanBoard/KanbanColumnList.module.scss
@@ -1,12 +1,3 @@
-.emptyContainer {
-  width: 100%;
-  height: 100%;
-}
-
-.emptyContent {
-  margin: 16px;
-}
-
 .columnList {
   display: flex;
   flex-direction: row;
