+++ src/templates/KanbanBoard/Layout.tsx
@@ -1,6 +1,7 @@
 import { AddStateModal } from "./AddStateModal";
 import { AddTaskModal } from "./AddTaskModal";
 import { EditTaskModal } from "./EditTaskModal";
+import { EmptyView } from "./EmptyView";
 import { KanbanColumnList } from "./KanbanColumnList";
 import styles from "./Layout.module.scss";
 import { Navbar } from "./Navbar";
@@ -15,15 +16,18 @@ export function Layout() {
 
       <div className={styles.content}>
         <KanbanColumnList
-          isStateDialogOpen={modals.addStateIsOpen}
-          onOpenStateDialog={() => dispatch({ key: "ADD_STATE::OPEN" })}
           onOpenEditTaskModal={(task) =>
             dispatch({ key: "EDIT_TASK::OPEN", payload: { task } })
           }
           onOpenCreateTaskModal={(stateId) =>
             dispatch({ key: "ADD_TASK::OPEN", payload: { stateId } })
           }
-        />
+        >
+          <EmptyView
+            onOpenStateDialog={() => dispatch({ key: "ADD_STATE::OPEN" })}
+            isStateDialogOpen={modals.addStateIsOpen}
+          />
+        </KanbanColumnList>
       </div>
 
       <AddStateModal
